+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ UNRELEASED
 
 * Dropped support for Sphinx < 3.0
 * Fixed ``TypeError`` when formatting a parametrized ``typing.IO`` annotation
+* Fixed data class displaying a return type in its ``__init__()`` method
 
 
 1.10.3